Splitting monolithic loaders file, consolidated data, changed to AsyncHtml
