--- 03_RAG/chunk_data/csv/gemini_generated.csv
@@ -1,21 +1,0 @@
-Name,Age,Favorite LLM
-Alice,25,Bard
-Bob,32,Megatron
-Charlie,18,Jurassic-1 Jumbo
-Diana,41,GPT-3
-Ethan,10,BLOOM
-Fiona,67,WuDao 2.0
-Gabriel,54,WuDao 2.0
-Hannah,29,Megatron
-Isaac,47,BARD
-Jasmine,38,Jurassic-1 Jumbo
-Kevin,22,WuDao 2.0
-Liam,15,BLOOM
-Mia,35,Megatron
-Noah,72,GPT-3
-Olivia,20,BARD
-Penelope,59,Jurassic-1 Jumbo
-Quentin,44,WuDao 2.0
-Ruby,31,BARD
-Samuel,19,Megatron
-Tina,63,GPT-3
